LPS-205650 Extract ScreenNavigationEntry methods from ServerPortalPropertiesScreenNavigationCategory to ServerPortalPropertiesScreenNavigationEntry and extends ServerPortalPropertiesScreenNavigationCategory in it
